backend/pyproject.toml
[project]
name = "netkitx-backend"
version = "0.1.0"
description = "NetKitX - Extensible Network Security Toolkit API"
requires-python = ">=3.11"

dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "asyncpg>=0.30.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "celery>=5.4.0",
    "redis>=5.0.0",
    "pyyaml>=6.0.0",
    "websockets>=13.0",
    "httpx>=0.27.0",
    "alembic>=1.14.0",
    "email-validator>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.27.0",
    "ruff>=0.8.0",
]

[tool.ruff]
target-version = "py311"
line-length = 100
